test/api/requirements.txt
Flask
requests_oauthlib
nose2